pyproject.toml
[project]
name = "cloudflare-particle"
version = "0.1.0"
description = "Update Cloudflare Access policy with Particle Cloud IPs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
